+++ go_expvar/CHANGELOG.md
@@ -1,6 +1,18 @@
 # CHANGELOG - go_expvar
 
-1.0.003-22-2017
+1.0.2 / Unreleased
+==================
+
+* [FEATURE] Make the hardcoded expvar URL path optional in the configuration
+
+1.0.1 / 5-22-2017
+==================
+
+### Changes
+
+* [FEATURE] ssl support for go_expvar check
+
+1.0.0 / 03-22-2017
 ==================
 
 ### Changes